requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# AI and ML
openai==1.3.7
langchain==0.0.340
langchain-openai==0.0.2
chromadb==0.4.18
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3

# Translation and NLP
googletrans==4.0.0rc1
deep-translator==1.11.4
spacy==3.7.2

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0